Updated emt-files script
